stream on boot things

+++ Coprocessor/cameraStreaming/nt_server.py
@@ -41,26 +41,29 @@ with cond:
 sd = NetworkTables.getTable("SmartDashboard")
 VISION_ERROR_CODE = -9999
 
-os.system('/home/pi/git/Mustang-Pi/cameraStreaming/mjpg_streamer_server.sh cam1 &')
+os.system('sudo python /home/pi/git/Mustang-Pi/cameraStreaming/watchdog.py "/home/pi/git/Mustang-Pi/cameraStreaming/mjpg_streamer_server.sh cam0 > /tmp/error0 2>&1" &')   
+os.system('sudo python /home/pi/git/Mustang-Pi/cameraStreaming/watchdog.py "/home/pi/git/Mustang-Pi/cameraStreaming/mjpg_streamer_server.sh cam1 > /tmp/error1 2>&1" &')   
 
 cam = '0'
 def valueChanged(table, key, value, isNew):
     global cam
-    if (key=='camera-source'):
+    if (key=='camera-source' and value=='next'):
         os.system('sudo killall mjpg_streamer')
         cam = int(cam)
         numCams = int(os.popen('ls -l /dev/ | egrep video.$ | wc -l').read().replace('\n', ''))
         cam = (cam + 1) % numCams
         cam = str(cam)
-        sd.putString('camera-source', '')
-        os.system('/home/pi/git/Mustang-Pi/cameraStreaming/mjpg_streamer_server.sh cam' + cam + ' &')	
+        sd.putString('camera-source', 'flipped')
 
 i = 0
 
-sd.addEntryListener(valueChanged)
+#sd.addEntryListener(valueChanged)
 
 while True:
     sd.putString('vision-status', "none")
     if (int(os.popen('ls -l /dev/ | egrep video.$ | wc -l').read().replace('\n', '')) == 0):
         sd.putString('warnings', 'no cameras found')
         sd.putString('vision-status', str(VISION_ERROR_CODE))
+    time.sleep(1)
+    if (os.system('grep "cleaning up resources" /tmp/error0') == 0):
+        os.system('sudo kill $(ps -aef | grep 8000 | grep mjpg_streamer | grep sudo | awk "{print $2}"')
